NavigationItem's Title now accepts non-ascii characters.  NavigationItem's External URL also now accepts non-ascii, but my understanding is that URLs only contain ascii so we replace any non-ascii characters.
globalnavigation_view wasn't displaying non-ascii characters correctly, so reverted back to previous code, but moved .decode('latin-1') in navigation_folders() so that top level folders display ok with non-ascii characters.

Found new bug that occurs when a second-level nav folder is retrieved and its parent has non-ascii characters -- since it converts non-ascii characters to "?" which errors out when used in searchResults() that is called to retrieve all sub folders for that parent.