Portlet assignment changes. From now on portlets for Quills objects (ones coming from Products.Quills.Weblog.Weblog) are assigned by-content-type using Generic Setup profile instead of relying on those set by quills.app.
For that purpose a portlets.xml file has been created in Products.Quills to enable a by-contenttype assignment of portlets.

I have also writen a migration script and registered it as a GS upgrade step to delete those by-interface assignments.

This will break the portlets assignments of QuillsEnabled objects present in the same Plone instance of a Quills instance, but we will be able to add the portlets through @@manage-portlets or through another upgrade step, and be able to modify the assigments.