Fix for Quills issue #172: The default view for WeblogEntries, i.e. WeblogEntryView, cannot be seen by appending '/view' to the entries URL. This is the way DynamicType allows access to the default view, and various default Plone views link to content (e.g. the Recent Items portlet).
Unfortunately this fix required different measures for Quills and 
QuillsEnabled. With the fix for the former I am quite satisfied: it uses 
the default DynamicType infrastructure now, and it's very plone'ish. The 
latter is, well, ugly. It hard-codes the view in the traversal machinery. 
While this is basicly what has been done before this patch to both Quills 
and QuillsEnabled, it's still bad. We need a way to work *with* 
DynamicType and FTI not against them!