* rolled back a few of the changes made in http://dev.plone.org/collective/changeset/109166/ZopeSkel/trunk
In general, we should err on the side of creating fully functioning templates for newer users over creating fully compliant code.  The reverted changes would have created broken templates for users who used zopeskel to create skeleton packages in line with current documentation.  We don't want that to happen, so we err in favor of asking more advanced developers to guard the pep8 and pyflakes friendliness of their code.