* added a new function 'getZipInfoForName' to resolve the potential differences between the identifying resource name extracted from a QTI manifest and the 'name' by which the corresponding file may be read from the ZipFile object being imported.
* refactored the 'readFromZip' function to rely on the new getZipInfoForName method to get the lookup value for reading the desired file from the ZipFile object.

both of these functions now rely on the 'unicodeDecode' python script in Products.ECQuiz.skins.ECQuiz to render lookup values and filenames into unicode strings for the sake of comparison.  This should mean that the results returned are onsistent leading to better matching at all stages of the process.

tests on qti import now work using both a zipfile with a manifest that matches the names in the zipfile information and one where the zipfile information has extra pathname information.